# OpenAPI.yml
openapi: 3.0.0
info:
  title: LumRPi API
  version: 1.0.0
servers:
  - url: http://lumrpi.local
paths:
  /status:
    get:
      tags:
        - default
      summary: Get status
      description: Récupère la config, les animations en cours sur les tables, etc.
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /flag:
    post:
      tags:
        - default
      summary: Jouer flag
      description: >-
        Joue un petit son et allume progressivement le bandeau de l'équipe en
        vert.


        Le paramètre table est un entier entre 1 et 18.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                table: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /box:
    post:
      tags:
        - default
      summary: Pwn box
      description: >-
        Jouer le son "Team number XX just pwned the box!", le bandeau passe en
        mode arc-en-ciel pour le reste du round.


        Le paramètre table est un entier entre 1 et 18.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                table: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /round:
    post:
      tags:
        - default
      summary: Démarrer un round
      description: >-
        - Jouer le son "Let's go for round number XX"
            
        - Le paramère round peut contenir un chiffre (1, 2 ou 3) ou le texte
        "final" pour la finale
            
        - Faire une animation de début de round sur le bandeau de la salle
            
        - Reset l'état de toutes les tables pour revenir à l'animation au repos
        (vague rose)
            
        - Prévenir (animation sur le bandeau de la salle + son) quand il reste 5
        min
            
        - Prévenir (animation sur le bandeau de la salle + son) quand il reste 1
        min
            
        - Le paramètre _duration_ est la durée du round en minutes, il est
        facultatif.
            
        - A la fin du round, jouer le son "The round is over" et le bandeau est
        freeze (plus de changement de pixel)
            

        Attention : Appeler cet endpoint planifie des animations dans le futur
        qui ne peuvent pas être annulées. Il est très fortement déconseillé de
        déclencher un nouveau round avant la fin du précédent.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                round: 1
                duration: 6
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /forceColor:
    post:
      tags:
        - default
      summary: Forcer couleur
      description: >-
        Forcer la couleur de toutes les tables (#000000 permet d'éteindre le
        bandeau). Si une durée (en minutes) est précisée, les animations
        reprennent une fois le temps écoulé (sauf si le round est terminé).


        En mode forcé, il n'y a plus aucune animation sur le bandeau mais il y a
        encore les animations sonores. Il faut démarrer un nouveau round sortir
        du mode forcé.


        Le but de cet endpoint est d'être appelé en dehors des rounds pour
        éclairer la salle.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                duration: 1
                color: '#000000'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}